api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.4
google-auth==2.38.0
requests==2.32.3
PyJWT==2.10.1
apscheduler==3.10.4
python-dateutil==2.9.0.post0
